Update TrafficSimulation2.py
Recent changes made it so that it no longer recognized bad input. This was due to a workaround for when it tried to find a 'type' attribute when vehicles did not have them. Updated it so that it will only find the type if it exists. This can be adjusted down the road if/when type is required. This same kind of approach can be used later if we decide to import optional attributes which may be helpful for calculations.

I also turned the errors into a list, they now print from the SimpleOutput.py which looks much cleaner. It may also come in useful later down the line so we can print it when and where we want in the GUI.