Fixed Program Errors
I have made adjustments to the following files:

AutomatiSimulation.py:
+ time library in order to use the sleep() and time() functions
+ created a while loop that ends at n time that invokes the simulation.update() function

TrafficSim2.xml:
+ changed a few values just for quicker testing purposes (specifically the cycles)

TrafficLightSimulation.py:
- removed green_light variable
+ added global dictionary that holds a road's traffic light and its green light value
+ changed a parameter
+ adjusted trafficLight_road, trafficLight_position, trafficLight_cycle syntax to its proper indices
+ time for each traffic light cycle is tracked and calculated correctly now
+ adjusted conditions to their correct syntax
+ added print statements to the console for debugging purposes

With these changes, it can run the program and simulate an output of the vehicles and traffic lights. You must run it through the AutomaticSimulation.py file in order to simulate.

The vehicle calculations are not 100% at the moment. It may have to do with the conditions within the TrafficLightSimulation.py file.
Any conerns, feel free to comment.