Rule #1 - Each vehicle must be on existing road
Added some checks to see if the vehicles are on a road. If not, removes them from the list. I also noticed that the error checking from previous commits no longer works. I changed the if statement for the crossroads to elif and I believe that resolved the issue. Tried testing with as many cases as possible but may have missed some edge cases. Let me know if something does not look right or you have any questions. Thanks!